Shorten the path of the "private" `isolated-hoist-non-react-statics-do-not-use-this-in-your-code`
