+++ .changeset/flat-schools-crash.md
@@ -1,0 +1,5 @@
+---
+'@emotion/react': patch
+---
+
+Shorten the path of the "private" `isolated-hoist-non-react-statics-do-not-use-this-in-your-code` entrypoint to avoid exeeding path limitations on Windows.
