+++ packages/react/_isolated-hnrs/package.json
@@ -1,0 +1,13 @@
+{
+  "main": "dist/emotion-react-_isolated-hnrs.cjs.js",
+  "module": "dist/emotion-react-_isolated-hnrs.esm.js",
+  "umd:main": "dist/emotion-react-_isolated-hnrs.umd.min.js",
+  "browser": {
+    "./dist/emotion-react-_isolated-hnrs.cjs.js": "./dist/emotion-react-_isolated-hnrs.browser.cjs.js",
+    "./dist/emotion-react-_isolated-hnrs.esm.js": "./dist/emotion-react-_isolated-hnrs.browser.esm.js"
+  },
+  "sideEffects": false,
+  "preconstruct": {
+    "umdName": "emotionHoistNonReactStatics"
+  }
+}
