--- packages/react/isolated-hoist-non-react-statics-do-not-use-this-in-your-code/package.json
@@ -1,13 +1,0 @@
-{
-  "main": "dist/emotion-react-isolated-hoist-non-react-statics-do-not-use-this-in-your-code.cjs.js",
-  "module": "dist/emotion-react-isolated-hoist-non-react-statics-do-not-use-this-in-your-code.esm.js",
-  "umd:main": "dist/emotion-react-isolated-hoist-non-react-statics-do-not-use-this-in-your-code.umd.min.js",
-  "browser": {
-    "./dist/emotion-react-isolated-hoist-non-react-statics-do-not-use-this-in-your-code.cjs.js": "./dist/emotion-react-isolated-hoist-non-react-statics-do-not-use-this-in-your-code.browser.cjs.js",
-    "./dist/emotion-react-isolated-hoist-non-react-statics-do-not-use-this-in-your-code.esm.js": "./dist/emotion-react-isolated-hoist-non-react-statics-do-not-use-this-in-your-code.browser.esm.js"
-  },
-  "sideEffects": false,
-  "preconstruct": {
-    "umdName": "emotionHoistNonReactStatics"
-  }
-}
