+++ packages/react/package.json
@@ -13,7 +13,7 @@
     "dist",
     "jsx-runtime",
     "jsx-dev-runtime",
-    "isolated-hoist-non-react-statics-do-not-use-this-in-your-code",
+    "_isolated-hnrs",
     "types/*.d.ts",
     "macro.js",
     "macro.d.ts",
@@ -68,7 +68,7 @@
       "./index.js",
       "./jsx-runtime.js",
       "./jsx-dev-runtime.js",
-      "./isolated-hoist-non-react-statics-do-not-use-this-in-your-code.js"
+      "./_isolated-hnrs.js"
     ],
     "umdName": "emotionReact"
   }
