+++ packages/react/src/utils.js
@@ -1,4 +1,4 @@
 // @flow
 export let isBrowser = typeof document !== 'undefined'
 
-export const hasOwnProperty = Object.prototype.hasOwnProperty
+export const hasOwnProperty = {}.hasOwnProperty
